refactor: modularize UI effects
